Update Program.cs
Added one line to set character encoding of Console to UTF-8 to allow UTF-8 characters be printed to Console correctly. This is important because other applications that are reading these values will be using them (such as camera name) to set settings for the camera, so all characters have to match with the camera name.